requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.41
botocore==1.29.41
charset-normalizer==2.1.1
discord.py==2.1.0
Django==4.1.4
django-cleanup==6.0.0
django-storages==1.13.2
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
multidict==6.0.4
python-dateutil==2.8.2
pytz==2022.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
yarl==1.8.2
